#4 first round of making django-tenants work
