+++ bakeup/core/admin.py
@@ -1,6 +1,8 @@
 from django.contrib import admin
 
-from bakeup.core.models import Address
+from django_tenants.admin import TenantAdminMixin
+
+from bakeup.core.models import Client
 
 # Register your models here.
 
@@ -9,6 +11,6 @@ class ExcludeAdminMixin(object):
     exclude = ('is_archived',)
 
 
-@admin.register(Address)
-class AddressAdmin(ExcludeAdminMixin, admin.ModelAdmin):
-    list_display = ('address',)
+@admin.register(Client)
+class ClientAdmin(TenantAdminMixin, admin.ModelAdmin):
+        list_display = ('name', 'created')
