+++ bakeup/shop/fixtures/demo_customers.json
@@ -1,1 +1,1 @@
-[{"model": "shop.customer", "pk": 1, "fields": {"created": "2022-04-26T13:13:57.606Z", "updated": "2022-04-26T13:13:57.606Z", "is_archived": false, "user": 4, "point_of_sale": 1}}, {"model": "shop.customer", "pk": 2, "fields": {"created": "2022-04-26T13:14:04.887Z", "updated": "2022-04-26T13:14:04.887Z", "is_archived": false, "user": 5, "point_of_sale": 1}}, {"model": "shop.customer", "pk": 3, "fields": {"created": "2022-05-11T08:46:51.535Z", "updated": "2022-05-11T08:46:51.535Z", "is_archived": false, "user": 1, "point_of_sale": 1}}, {"model": "shop.pointofsale", "pk": 1, "fields": {"created": "2022-04-26T13:13:55.350Z", "updated": "2022-04-26T13:13:55.350Z", "is_archived": false, "name": "Pumpwerk", "address": 1}}, {"model": "core.address", "pk": 1, "fields": {"address": "Am Pumpwerk in Grünau"}}]
+[{"model": "shop.customer", "pk": 1, "fields": {"created": "2022-04-26T13:13:57.606Z", "updated": "2022-04-26T13:13:57.606Z", "is_archived": false, "user": 4, "point_of_sale": 1}}, {"model": "shop.customer", "pk": 2, "fields": {"created": "2022-04-26T13:14:04.887Z", "updated": "2022-04-26T13:14:04.887Z", "is_archived": false, "user": 5, "point_of_sale": 1}}, {"model": "shop.customer", "pk": 3, "fields": {"created": "2022-05-11T08:46:51.535Z", "updated": "2022-05-11T08:46:51.535Z", "is_archived": false, "user": 1, "point_of_sale": 1}}, {"model": "shop.pointofsale", "pk": 1, "fields": {"created": "2022-04-26T13:13:55.350Z", "updated": "2022-04-26T13:13:55.350Z", "is_archived": false, "name": "Pumpwerk", "address": 1}}, {"model": "contrib.Address", "pk": 1, "fields": {"address": "Am Pumpwerk in Grünau"}}]
