+++ bakeup/templates/workshop/includes/sidebar.html
@@ -2,6 +2,7 @@
 <div class="col-auto col-md-3 col-xl-2 px-sm-2 px-0 bg-dark d-print-none">
   <nav class="d-flex flex-column align-items-center align-items-sm-start min-vh-100">
       {% include "includes/logo.html" %}
+      <h2 class="mt-0 mb-0 text-center text-light w-100 text-uppercase">{{ request.tenant.name }}</h2>
         <button class="navbar-toggler navbar-toggler-right" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
           <span class="navbar-toggler-icon"></span>
         </button>
