+++ requirements/base.txt
@@ -20,3 +20,4 @@ django-treebeard
 sorl-thumbnail
 django_bootstrap5
 django_filter
+django-tenants
